product_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0
